Conflicts:
	src/core-impl/collections/db/sql/SqlMeta.cpp
	src/core-impl/collections/db/sql/SqlScanResultProcessor.cpp
	src/core-impl/meta/file/File.cpp
	utilities/collectionscanner/Track.cpp

+++ src/covermanager/CoverViewDialog.h
@@ -31,7 +31,7 @@ class AMAROK_EXPORT CoverViewDialog : public QDialog
 
     private slots:
         void updateCaption();
-        void zoomFactorChanged( float value );
+        void zoomFactorChanged( qreal value );
 
     private:
         void createViewer( const QPixmap &pixmap, const QWidget *widget );
